Refactor Category Chip
